Fix bug for mp4 in h5 (#882)
* fix h5 for timestamp change

* fix bug for mp4 in h5

+++ funasr/runtime/html5/static/main.js
@@ -110,7 +110,7 @@ function play_file()
 }
 function start_file_send()
 {
-		sampleBuf=new Int16Array( file_data_array );
+		sampleBuf=new Uint8Array( file_data_array );
  
 		var chunk_size=960; // for asr chunk_size [5, 10, 5]
